+++ test/smoke_test.dart
@@ -1,0 +1,4 @@
+// TODO Go crazy.
+void main(){
+
+}